elk/install/flask-hello/requirements.txt
anyio==4.0.0
blinker==1.7.0
certifi==2023.7.22
click==8.1.7
Flask==3.0.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-telegram-bot==20.6
sniffio==1.3.0
Werkzeug==3.0.1
